// jobup-frontend/src/app/generated-sources/openapi/model/workerUpdateDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * DTO used to update an existing worker
 */
export interface WorkerUpdateDto { 
    fullName?: string;
    jobType?: string;
    location?: string;
    phoneNumber?: string;
    description?: string;
}